Create controllers and methods http

+++ src/main/java/com/api/alunos/repositories/AlunosRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.api.alunos.entities.Alunos;
 
-public interface AlunosRepository extends JpaRepository<Alunos, Long>{
+public interface AlunosRepository extends JpaRepository<Alunos, Integer>{
 }
